feat:实现接口幂等，完善限流

+++ acimage_community/src/main/java/com/acimage/community/service/star/impl/StarWriteServiceImpl.java
@@ -7,6 +7,7 @@ import com.acimage.common.global.exception.BusinessException;
 import com.acimage.common.model.domain.community.Star;
 import com.acimage.community.service.cmtyuser.CmtyUserWriteService;
 import com.acimage.community.global.consts.StarKeyConstants;
+import com.acimage.community.service.star.StarQueryService;
 import com.acimage.community.service.topic.TopicQueryService;
 import com.acimage.community.service.star.StarWriteService;
 import com.acimage.common.utils.redis.RedisUtils;
@@ -18,6 +19,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.stereotype.Service;
 
+import java.util.Date;
 import java.util.concurrent.TimeUnit;
 
 @Slf4j
@@ -26,6 +28,8 @@ public class StarWriteServiceImpl implements StarWriteService {
 
     @Autowired
     StarDao starDao;
+    @Autowired
+    StarQueryService starQueryService;
     @Autowired
     RedisUtils redisUtils;
     @Autowired
@@ -37,7 +41,15 @@ public class StarWriteServiceImpl implements StarWriteService {
 
     @Override
     public void saveStar(long userId, long topicId) {
+        if (starQueryService.isStar(userId, topicId)) {
+            log.info("user:{}已经点赞过话题topicId:{}", UserContext.getUsername(), topicId);
+            throw new BusinessException("已经点过赞了");
+        }
+
         Star star = new Star(userId, topicId);
+        star.setUserId(userId);
+        star.setTopicId(topicId);
+        star.setCreateTime(new Date());
 
         try {
             starDao.insert(star);
@@ -46,18 +58,24 @@ public class StarWriteServiceImpl implements StarWriteService {
             throw new BusinessException("已经收藏过了，请刷新重试");
         }
 
-        redisUtils.setObjectJson(StarKeyConstants.keyOfIsStar(userId, topicId), Boolean.TRUE, 30, TimeUnit.SECONDS);
-
-        int increment=1;
+        int increment = 1;
         topicSpAttrWriteService.increaseStarCount(topicId, increment);
+        //更新主人star
         Topic topic = topicQueryService.getTopic(topicId);
         if (topic != null) {
-            cmtyUserWriteService.updateStarCountByIncrement(topic.getUserId(),increment);
+            cmtyUserWriteService.updateStarCountByIncrement(topic.getUserId(), increment);
         }
+
+        long timeout = 3L;
+        redisUtils.setObjectJson(StarKeyConstants.keyOfIsStar(userId, topicId), Boolean.TRUE, timeout, TimeUnit.SECONDS);
     }
 
     @Override
     public void removeStar(long userId, long topicId) {
+        if (!starQueryService.isStar(userId, topicId)) {
+            log.info("user:{}已经取消点赞过话题topicId:{}", UserContext.getUsername(), topicId);
+            throw new BusinessException("已经取消过点赞了");
+        }
         int col = starDao.deleteByUserIdAndTopicId(userId, topicId);
         //删除缓存
         redisUtils.delete(StarKeyConstants.keyOfIsStar(userId, topicId));
